services/indexing-service/requirements.txt
fastapi==0.95.1
uvicorn==0.22.0
pydantic==1.10.7
dapr-client==1.9.0
ragatouille==0.0.8
numpy==1.24.3
